Update server.cfg
There are a few reasons for this change.

Initially, we wanted to make it easier for you to find the cvars you need to change, but we found this causes confusion. With a smaller file, it's less overwhelming for server operators to keep track of what is actually set on their servers.